fix(ui):Renamed field directory
